[MOV] website_sale: extract cart, checkout & address management
Move the code whose effect and scope is clearly known to be
limited to a specific page to dedicated js files.

Incoming tasks will progressively try to split/move the rest of
the js logic to place it in the right files when we are sure
of its uses.

Part-of: odoo/odoo#144429
