+++ addons/website_sale/__manifest__.py
@@ -64,6 +64,9 @@
             'website_sale/static/src/scss/website_sale_delivery.scss',
             'website/static/lib/multirange/multirange_custom.scss',
             'sale/static/src/scss/sale_portal.scss',
+            'website_sale/static/src/js/address.js',
+            'website_sale/static/src/js/cart.js',
+            'website_sale/static/src/js/checkout.js',
             'website_sale/static/src/js/payment_button.js',
             'website_sale/static/src/js/payment_form.js',
             'website_sale/static/src/js/sale_variant_mixin.js',
